app/python/requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile pyproject.toml -o app/python/requirements.txt
annotated-types==0.7.0
    # via pydantic
faker==37.12.0
    # via fds-aws-coding-exercise (pyproject.toml)
pydantic==2.12.4
    # via fds-aws-coding-exercise (pyproject.toml)
pydantic-core==2.41.5
    # via pydantic
python-dotenv==1.2.1
    # via fds-aws-coding-exercise (pyproject.toml)
typing-extensions==4.15.0
    # via
    #   pydantic
    #   pydantic-core
    #   typing-inspection
typing-inspection==0.4.2
    # via pydantic
tzdata==2025.2
    # via faker
